BarcodeGenerator/requirements.txt
Pillow
barcode
pyyaml
pywin32
